+++ CHANGELOG.md
@@ -82,6 +82,7 @@
   * added `stubbles\lang\reflect\annotation\Annotation::targetName()` which returns the name of the class, method, function, property or parameter of where the annotation comes from
   * added `stubbles\peer\http\Http::RFC_2616` and `stubbles\peer\http\Http::RFC_7230`
   * added `stubbles\peer\http\Http::isValidRfc()`
+  * added `stubbles\peer\http\Http::lines()`
   * added `stubbles\peer\http\Http::OPTIONS`
   * added `stubbles\peer\http\HttpVersion`
   * added `stubbles\peer\http\HttpUri::castFrom()`